requirements.txt
azure-search-documents==11.4.0
azure-core==1.29.5
python-dotenv==1.0.0
